Hoist CommCalculation to its own file
Thus completes the Big Break Up at step 5.
