+++ PyRoute/Galaxy.py
@@ -14,7 +14,8 @@ import math
 import networkx as nx
 
 from PyRoute.Star import Star
-from PyRoute.Calculation.TradeCalculation import TradeCalculation, CommCalculation
+from PyRoute.Calculation.TradeCalculation import TradeCalculation
+from PyRoute.Calculation.CommCalculation import CommCalculation
 from PyRoute.Calculation.OwnedWorldCalculation import OwnedWorldCalculation
 from PyRoute.Calculation.NoneCalculation import NoneCalculation
 from PyRoute.Calculation.XRouteCalculation import XRouteCalculation
